pyproject.toml
[tool.poetry]
name = "backuper-nq"  # just for PyPI
version = "0.2.0"
description = "Backup tool with CI-like configuration"
authors = ["niqzart <niqzart@gmail.com>"]
readme = "README.md"
packages = [{ include = "backuper" }]

[tool.poetry.scripts]
backuper = "backuper.__main__:cli"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.12.3"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
flake8 = "7.0.0"
black = "24.4.2"
pre-commit = "^3.2.1"
mypy = "^1.1.1"
flake8-pie = "0.16.0"
dlint = "0.14.0"
flake8-coding = "1.3.2"
flake8-secure-coding-standard = "1.4.1"
flake8-unused-arguments = "0.0.13"
flake8-pytest-style = "1.7.2"
flake8-useless-assert = "0.4.4"
flake8-simplify = "0.19.3"
flake8-future-import = "0.4.7"
flake8-functions-names = "0.4.0"
flake8-variables-names = "0.0.5"
flake8-comments = "0.1.2"
flake8-annotations = "2.9.1"
flake8-new-union-types = "0.4.1"
flake8-pep585 = "0.1.7"
flake8-type-checking = "2.3.1"
flake8-print = "5.0.0"
flake8-use-fstring = "1.4"
flake8-use-pathlib = "0.3.0"
flake8-test-name = "^0.1.5"
wemake-python-styleguide = "0.18.0"
types-pyyaml = "^6.0.12.20240311"

[tool.isort]
profile = "black"
py_version = 312
combine_as_imports = true
known_first_party = ["app", "tests"]
no_lines_before = "LOCALFOLDER"
reverse_relative = true
line_length = 88

[tool.mypy]
strict = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
